Added files via upload
VREData.csv is the file Guy placed on the County API.
VREDateRow1 is the file that I (misspelled and) started to manually parse out the questions from the data file. I have about 82 lines of data and have only about 40% of one site finished. I wanted this as a sample, so that we can see what to expect from the County Data in the future. I think I will have to work on doing a real parsing at this point, probably with a file stream.